% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-distribution.R
\name{build_dist_plot}
\alias{build_dist_plot}
\title{Single-variable distribution plot}
\usage{
build_dist_plot(df, var, opts = dist_options())
}
\arguments{
\item{df}{The working data frame.}

\item{var}{Name of the numeric column to plot.}

\item{opts}{Options as built by [dist_options()].}
}
\value{
A ggplot object; a [message_plot()] when the variable cannot be
  plotted.
}
\description{
Single-variable distribution plot
}
\examples{
build_dist_plot(mtcars, "mpg", dist_options(plot_type = "density"))
}
