% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-correlation.R
\name{corr_export_table}
\alias{corr_export_table}
\title{Flatten the correlation and p-value matrices into one exportable table}
\usage{
corr_export_table(obj)
}
\arguments{
\item{obj}{A list as produced by [corr_objects()].}
}
\value{
A long data frame with a `Statistic` column marking `r` and `p`
  blocks.
}
\description{
Flatten the correlation and p-value matrices into one exportable table
}
