% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-correlation.R
\name{corr_method_choices}
\alias{corr_method_choices}
\title{Correlation methods offered on the Correlations tab}
\usage{
corr_method_choices()
}
\value{
Named character vector for `selectInput(choices = )`.
}
\description{
Correlation methods offered on the Correlations tab
}
