% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-distribution.R
\name{dist_options}
\alias{dist_options}
\title{Default options for the distribution plots}
\usage{
dist_options(...)
}
\arguments{
\item{...}{Values overriding the defaults.}
}
\value{
A list of plot options.
}
\description{
Default options for the distribution plots
}
