% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-distribution.R
\name{dist_plot_choices}
\alias{dist_plot_choices}
\title{Distribution plot types}
\usage{
dist_plot_choices()
}
\value{
Named character vector for `radioButtons(choices = )`.
}
\description{
Distribution plot types
}
