% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expand_palette}
\alias{expand_palette}
\title{Expand a colour palette to an arbitrary number of levels}
\usage{
expand_palette(colors, n)
}
\arguments{
\item{colors}{Character vector of colours.}

\item{n}{Number of colours required.}
}
\value{
Character vector of length `n`.
}
\description{
Grouping variables often have more levels than a journal palette has
colours; this interpolates rather than recycling, which would produce
duplicate group colours.
}
\examples{
expand_palette(c("#E64B35", "#4DBBD5"), 5)
}
