% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_pkg}
\alias{has_pkg}
\title{Is an optional package available?}
\usage{
has_pkg(pkg)
}
\arguments{
\item{pkg}{Package name.}
}
\value{
`TRUE` when the package can be loaded.
}
\description{
Is an optional package available?
}
