% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{plot_theme_choices}
\alias{plot_theme_choices}
\title{Available plot themes}
\usage{
plot_theme_choices()
}
\value{
Named character vector suitable for `selectInput(choices = )`.
}
\description{
Available plot themes
}
