% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regression_fit_stats}
\alias{regression_fit_stats}
\title{Printable fit statistics for a linear model}
\usage{
regression_fit_stats(rf)
}
\arguments{
\item{rf}{A successful result from [fit_regression()].}
}
\value{
A character vector, one line per statistic.
}
\description{
Printable fit statistics for a linear model
}
