% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{save_plot}
\alias{save_plot}
\title{Save a ggplot using the current export settings}
\usage{
save_plot(plot, file, export = default_export())
}
\arguments{
\item{plot}{A ggplot (or patchwork / grob) object.}

\item{file}{Destination path.}

\item{export}{A list as produced by [default_export()].}
}
\value{
Invisibly, `file`.
}
\description{
Save a ggplot using the current export settings
}
