% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots-correlation.R
\name{scatter_matrix_plot}
\alias{scatter_matrix_plot}
\title{Scatter plot matrix}
\usage{
scatter_matrix_plot(
  df,
  colors = get_palette("Nature"),
  gg_theme = build_plot_theme(),
  font_size = 13
)
}
\arguments{
\item{df}{A data frame of numeric columns.}

\item{colors}{A colour palette, see [get_palette()].}

\item{gg_theme}{A ggplot2 theme, see [build_plot_theme()].}

\item{font_size}{Base font size.}
}
\value{
A `ggmatrix` object.
}
\description{
Scatter plot matrix
}
