% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_choices}
\alias{transform_choices}
\title{Transformations offered on the Data Prep tab}
\usage{
transform_choices()
}
\value{
Named character vector for `selectInput(choices = )`.
}
\description{
Transformations offered on the Data Prep tab
}
